requirements.txt
sounddevice==0.4.6
scipy
openai
faster-whisper==0.10.0
pydub
python-dotenv==1.0.1
ttkbootstrap
